requirements.txt
# MeetingMind Analysis API — Dependencies

# ── Web Framework ──────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic[email]>=2.0.0

# ── Database ───────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9         # PostgreSQL driver

# ── Authentication ─────────────────────────────────────────────────────────────
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# ── Background Jobs ────────────────────────────────────────────────────────────
apscheduler>=3.10.0

# ── Audio Processing ───────────────────────────────────────────────────────────
ffmpeg-python>=0.2.0

# ── Speech-to-Text ─────────────────────────────────────────────────────────────
faster-whisper>=1.0.0

# ── Speaker Diarization ────────────────────────────────────────────────────────
pyannote.audio>=3.1.0
torchaudio>=2.0.0

# ── LLM Summarization ──────────────────────────────────────────────────────────
groq>=0.9.0
anthropic>=0.25.0

# ── Utilities ──────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
aiofiles>=23.0.0